Move calico manifest handling to createClusterctlLocalRepository func()
