+++ test/e2e/e2e_suite_test.go
@@ -16,7 +16,10 @@ import (
 	. "github.com/onsi/gomega"
 	"k8s.io/apimachinery/pkg/runtime"
 
+	"github.com/jinzhu/copier"
+	"gopkg.in/yaml.v3"
 	capi_e2e "sigs.k8s.io/cluster-api/test/e2e"
+
 	"sigs.k8s.io/cluster-api/test/framework"
 	"sigs.k8s.io/cluster-api/test/framework/bootstrap"
 	"sigs.k8s.io/cluster-api/test/framework/clusterctl"
@@ -162,6 +165,11 @@ func createClusterctlLocalRepository(config *clusterctl.E2EConfig, repositoryFol
 	// Ensuring a CNI file is defined in the config and register a FileTransformation to inject the referenced file as in place of the CNI_RESOURCES envSubst variable.
 	Expect(config.Variables).To(HaveKey(capi_e2e.CNIPath), "Missing %s variable in the config", capi_e2e.CNIPath)
 	cniPath := config.GetVariable(capi_e2e.CNIPath)
+	if osType == "centos" {
+		updateCalico(cniPath, "eth1")
+	} else {
+		updateCalico(cniPath, "enp2s0")
+	}
 	Expect(cniPath).To(BeAnExistingFile(), "The %s variable should resolve to an existing file", capi_e2e.CNIPath)
 	createRepositoryInput.RegisterClusterResourceSetConfigMapTransformation(cniPath, capi_e2e.CNIResources)
 
@@ -224,4 +232,42 @@ func validateGlobals(specName string) {
 	Expect(bootstrapClusterProxy).ToNot(BeNil(), "Invalid argument. bootstrapClusterProxy can't be nil when calling %s spec", specName)
 	Expect(osType).ToNot(Equal(""))
 	Expect(os.MkdirAll(artifactFolder, 0755)).To(Succeed(), "Invalid argument. artifactFolder can't be created for %s spec", specName)
+}
+
+func updateCalico(calicoYaml, calicoInterface string) {
+	err := downloadFile(calicoYaml, "https://docs.projectcalico.org/manifests/calico.yaml")
+	Expect(err).To(BeNil(), "Unable to download Calico manifest")
+	cniYaml, err := os.ReadFile(calicoYaml)
+	Expect(err).To(BeNil(), "Unable to read Calico manifest")
+	podCIDR := os.Getenv("POD_CIDR")
+	cniYaml = []byte(strings.Replace(string(cniYaml), "192.168.0.0/16", podCIDR, -1))
+
+	yamlDocuments, err := splitYAML(cniYaml)
+	Expect(err).To(BeNil(), "Cannot unmarshal the calico yaml elements to golang objects")
+	calicoNodes, err := yamlContainKeyValue(yamlDocuments, "calico-node", "metadata", "labels", "k8s-app")
+	Expect(err).To(BeNil())
+	for _, calicoNode := range calicoNodes {
+		calicoNodeSpecTemplateSpec, err := yamlFindByValue(calicoNode, "spec", "template", "spec", "containers")
+		Expect(err).To(BeNil())
+		calicoNodeContainers, err := yamlContainKeyValue(calicoNodeSpecTemplateSpec.Content, "calico-node", "name")
+		Expect(err).To(BeNil())
+		// Since we find the container by name, we expect to get only one container.
+		Expect(len(calicoNodeContainers) == 1).To(BeTrue(), "Found 0 or more than 1 container with name `calico-node`")
+		calicoNodeContainer := calicoNodeContainers[0]
+		calicoNodeContainerEnvs, err := yamlFindByValue(calicoNodeContainer, "env")
+		Expect(err).To(BeNil())
+		addItem := &yaml.Node{}
+		err = copier.CopyWithOption(addItem, calicoNodeContainerEnvs.Content[0], copier.Option{IgnoreEmpty: true, DeepCopy: true})
+		Expect(err).To(BeNil(), "Cannot copy this object")
+		addItem.Content[1].SetString("IP_AUTODETECTION_METHOD")
+		addItem.Content[3].SetString("interface=" + calicoInterface)
+		addItem.HeadComment = "Start section modified by CAPM3 e2e test framework"
+		addItem.FootComment = "End section modified by CAPM3 e2e test framework"
+		calicoNodeContainerEnvs.Content = append(calicoNodeContainerEnvs.Content, addItem)
+	}
+
+	yamlOut, err := printYaml(yamlDocuments)
+	Expect(err).To(BeNil())
+	err = os.WriteFile(calicoYaml, yamlOut, 0664)
+	Expect(err).To(BeNil(), "Cannot print out the update to the file")
 }
