+++ test/e2e/e2e_test.go
@@ -6,13 +6,12 @@ import (
 	"path/filepath"
 	"strings"
 
-	"github.com/jinzhu/copier"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
-	"gopkg.in/yaml.v3"
 
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	capi "sigs.k8s.io/cluster-api/api/v1beta1"
+	capi_e2e "sigs.k8s.io/cluster-api/test/e2e"
 	"sigs.k8s.io/cluster-api/test/framework"
 	"sigs.k8s.io/cluster-api/test/framework/clusterctl"
 )
@@ -25,7 +24,6 @@ var (
 	cluster                  *capi.Cluster
 	clusterName              = "test1"
 	clusterctlLogFolder      string
-	cniFile                  string
 	controlplaneListOptions  metav1.ListOptions
 	targetCluster            framework.ClusterProxy
 	controlPlaneMachineCount int64
@@ -34,16 +32,10 @@ var (
 var _ = Describe("Workload cluster creation", func() {
 
 	BeforeEach(func() {
-		cniFile = "/tmp/calico.yaml"
 		osType := strings.ToLower(os.Getenv("OS"))
 		Expect(osType).ToNot(Equal(""))
 		flavorSuffix = "-m3-dev-env"
 
-		if osType == "centos" {
-			updateCalico(cniFile, "eth1")
-		} else {
-			updateCalico(cniFile, "enp2s0")
-		}
 		validateGlobals(specName)
 
 		// We need to override clusterctl apply log folder to avoid getting our credentials exposed.
@@ -74,7 +66,7 @@ var _ = Describe("Workload cluster creation", func() {
 					ControlPlaneMachineCount: &controlPlaneMachineCount,
 					WorkerMachineCount:       &workerMachineCount,
 				},
-				CNIManifestPath:              cniFile,
+				CNIManifestPath:              e2eConfig.GetVariable(capi_e2e.CNIPath),
 				WaitForClusterIntervals:      e2eConfig.GetIntervals(specName, "wait-cluster"),
 				WaitForControlPlaneIntervals: e2eConfig.GetIntervals(specName, "wait-control-plane"),
 				WaitForMachineDeployments:    e2eConfig.GetIntervals(specName, "wait-worker-nodes"),
@@ -90,41 +82,3 @@ var _ = Describe("Workload cluster creation", func() {
 		})
 	})
 })
-
-func updateCalico(calicoYaml, calicoInterface string) {
-	err := downloadFile(calicoYaml, "https://docs.projectcalico.org/manifests/calico.yaml")
-	Expect(err).To(BeNil(), "Unable to download Calico manifest")
-	cniYaml, err := os.ReadFile(calicoYaml)
-	Expect(err).To(BeNil(), "Unable to read Calico manifest")
-	podCIDR := os.Getenv("POD_CIDR")
-	cniYaml = []byte(strings.Replace(string(cniYaml), "192.168.0.0/16", podCIDR, -1))
-
-	yamlDocuments, err := splitYAML(cniYaml)
-	Expect(err).To(BeNil(), "Cannot unmarshal the calico yaml elements to golang objects")
-	calicoNodes, err := yamlContainKeyValue(yamlDocuments, "calico-node", "metadata", "labels", "k8s-app")
-	Expect(err).To(BeNil())
-	for _, calicoNode := range calicoNodes {
-		calicoNodeSpecTemplateSpec, err := yamlFindByValue(calicoNode, "spec", "template", "spec", "containers")
-		Expect(err).To(BeNil())
-		calicoNodeContainers, err := yamlContainKeyValue(calicoNodeSpecTemplateSpec.Content, "calico-node", "name")
-		Expect(err).To(BeNil())
-		// Since we find the container by name, we expect to get only one container.
-		Expect(len(calicoNodeContainers) == 1).To(BeTrue(), "Found 0 or more than 1 container with name `calico-node`")
-		calicoNodeContainer := calicoNodeContainers[0]
-		calicoNodeContainerEnvs, err := yamlFindByValue(calicoNodeContainer, "env")
-		Expect(err).To(BeNil())
-		addItem := &yaml.Node{}
-		err = copier.CopyWithOption(addItem, calicoNodeContainerEnvs.Content[0], copier.Option{IgnoreEmpty: true, DeepCopy: true})
-		Expect(err).To(BeNil(), "Cannot copy this object")
-		addItem.Content[1].SetString("IP_AUTODETECTION_METHOD")
-		addItem.Content[3].SetString("interface=" + calicoInterface)
-		addItem.HeadComment = "Start section modified by CAPM3 e2e test framework"
-		addItem.FootComment = "End section modified by CAPM3 e2e test framework"
-		calicoNodeContainerEnvs.Content = append(calicoNodeContainerEnvs.Content, addItem)
-	}
-
-	yamlOut, err := printYaml(yamlDocuments)
-	Expect(err).To(BeNil())
-	err = os.WriteFile(calicoYaml, yamlOut, 0664)
-	Expect(err).To(BeNil(), "Cannot print out the update to the file")
-}
